SOLR-14408 Refactor MoreLikeThisHandler implementation
